@@ -1,6 +1,6 @@
 {
   "name": "parse-server-azure-storage",
-  "version": "1.1.0",
+  "version": "1.1.1",
   "description": "Use Azure Blob Storage with Parse Server",
   "main": "lib/AzureStorageAdapter.js",
   "repository": {
@@ -14,26 +14,22 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "azure-storage": "^1.0.1",
-    "babel-polyfill": "^6.5.0",
-    "babel-runtime": "^6.5.0"
+    "@azure/storage-blob": "^12.17.0",
+    "core-js": "^3.35.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.5.1",
-    "babel-core": "^6.5.1",
-    "babel-istanbul": "^0.7.0",
-    "babel-preset-es2015": "^6.5.0",
-    "babel-preset-stage-0": "^6.5.0",
-    "babel-register": "^6.5.1",
-    "codecov": "^1.0.1",
-    "jasmine": "^2.4.1",
-    "parse-server-conformance-tests": "^1.0.0"
+    "@babel/cli": "^7.23.4",
+    "@babel/core": "^7.23.4",
+    "@babel/preset-env": "^7.23.4",
+    "@babel/register": "^7.22.15",
+    "jest": "^29.7.0",
+    "dotenv": "^16.3.1" 
   },
   "scripts": {
-    "build": "./node_modules/.bin/babel src/ -d lib/",
-    "test": "./node_modules/.bin/babel-node ./node_modules/babel-istanbul/lib/cli.js cover -x **/spec/** ./node_modules/jasmine/bin/jasmine.js"
+    "build": "babel src/ -d lib/",
+    "test": "jest"
   },
   "engines": {
-    "node": ">=4.3"
+    "node": ">=14"
   }
 }
