@@ -2,7 +2,7 @@
 //
 // Stores Parse files in Azure Blob Storage.
 
-import * as Azure from 'azure-storage';
+import { BlobServiceClient, PublicAccessType } from '@azure/storage-blob';
 import requiredParameter from './RequiredParameter';
 
 export class AzureStorageAdapter {
@@ -19,77 +19,75 @@ export class AzureStorageAdapter {
     this._directAccess = directAccess;
 
     // Init client
-    this._client = Azure.createBlobService(this._accountName, this._accessKey);
+    const connectionString = `DefaultEndpointsProtocol=https;AccountName=${accountName};AccountKey=${accessKey};EndpointSuffix=core.windows.net`;
+    this._client = BlobServiceClient.fromConnectionString(connectionString);
+    this._containerClient = this._client.getContainerClient(container);
   }
 
   /**
-   * For a given config object, filename, and data, store a file in Azure Blob Storage
-   * @param  {object} config
+   * For a given filename and data, store a file in Azure Blob Storage
    * @param  {string} filename
    * @param  {string} data
    * @return {Promise} Promise containing the Azure Blob Storage blob creation response
    */
-  createFile(filename, data) {
-    let containerParams = {
-      publicAccessLevel: (this._directAccess) ? 'blob' : undefined
-    };
+  async createFile(filename, data) {
+    try {
+      // Create container if it doesn't exist with proper access level
+      const options = this._directAccess ? { access: 'blob' } : {};
+      await this._containerClient.createIfNotExists(options);
 
-    return new Promise((resolve, reject) => {
-      this._client.createContainerIfNotExists(this._container, containerParams, (cerr, cresult, cresponse) => {
-        if (cerr) {
-          return reject(cerr);
-        }
-
-        this._client.createBlockBlobFromText(this._container, filename, data, (err, result) => {
-          if (err) {
-            return reject(err);
-          }
-
-          resolve(result);
-        });
-      });
-    });
+      // Upload the file
+      const blockBlobClient = this._containerClient.getBlockBlobClient(filename);
+      const buffer = Buffer.from(data);
+      const uploadResponse = await blockBlobClient.upload(buffer, buffer.length);
+      return uploadResponse;
+    } catch (error) {
+      throw new Error(`Error creating file: ${error.message}`);
+    }
   }
 
   /**
    * Delete a file if found by filename
-   * @param  {object} config
    * @param  {string} filename
    * @return {Promise} Promise that succeeds with the result from Azure Storage
    */
-  deleteFile(filename) {
-    return new Promise((resolve, reject) => {
-      this._client.deleteBlob(this._container, filename, (err, res) => {
-          if (err) {
-            return reject(err);
-          }
-
-          resolve(res);
-        });
-    });
+  async deleteFile(filename) {
+    try {
+      const blockBlobClient = this._containerClient.getBlockBlobClient(filename);
+      const deleteResponse = await blockBlobClient.delete();
+      return deleteResponse;
+    } catch (error) {
+      throw new Error(`Error deleting file: ${error.message}`);
+    }
   }
 
   /**
    * Search for and return a file if found by filename
-   * @param  {object} config
    * @param  {string} filename
-   * @return {Promise} Promise that succeeds with the result from Azure Storage
+   * @return {Promise<Buffer>} Promise that succeeds with the file data as a Buffer
    */
-  getFileData(filename) {
-    return new Promise((resolve, reject) => {
-      this._client.getBlobToText(this._container, filename, (err, text, blob, res) => {
-        if (err) {
-          return reject(err);
-        }
-
-        resolve(new Buffer(text));
+  async getFileData(filename) {
+    try {
+      const blockBlobClient = this._containerClient.getBlockBlobClient(filename);
+      const downloadResponse = await blockBlobClient.download(0);
+      
+      return new Promise((resolve, reject) => {
+        const chunks = [];
+        downloadResponse.readableStreamBody.on('data', (chunk) => {
+          chunks.push(chunk instanceof Buffer ? chunk : Buffer.from(chunk));
+        });
+        downloadResponse.readableStreamBody.on('end', () => {
+          resolve(Buffer.concat(chunks));
+        });
+        downloadResponse.readableStreamBody.on('error', reject);
       });
-    });
+    } catch (error) {
+      throw new Error(`Error getting file data: ${error.message}`);
+    }
   }
 
   /**
-   * Generates and returns the location of a file stored in Azure Blob Storage for the given request and filename
-   * The location is the direct Azure Blob Storage link if the option is set, otherwise we serve the file through parse-server
+   * Generates and returns the location of a file stored in Azure Blob Storage
    * @param  {object} config
    * @param  {string} filename
    * @return {string} file's url
